+++ src/App.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes,Navigate } from 'react-router-dom';
 import Registration from './registration-layout/Registration'; // ваш компонент регистрации
 import LoginForm from './login-layout/LoginForm';
-import AddEventPage from "./Event/event-add-edit-layout/AddEventPage";
-import EventsPage from "./Event/events-layout/EventsPage";
-import ActivitiesPage from "./Activity/activities-layout/ActivitiesPage";
-import AddActivityPage from "./Activity/activities-layout/add-edit-activity-layout/AddActivityPage";
-import EditEventPage from "./Event/event-add-edit-layout/EditEventPage";
-import EditActivityPage from "./Activity/activities-layout/add-edit-activity-layout/EditActivityPage";
-import EditUserProfilePage from "./User/user-profile-layout/EditUserProfile";
-import UserRequestsPage from "./User/user-requests-layout/UserRequestsPage";
+import AddEventPage from "./event/event-add-edit-layout/AddEventPage";
+import EventsPage from "./event/events-layout/EventsPage";
+import ActivitiesPage from "./activity/activities-layout/ActivitiesPage";
+import AddActivityPage from "./activity/add-edit-activity-layout/AddActivityPage";
+import EditEventPage from "./event/event-add-edit-layout/EditEventPage";
+import EditActivityPage from "./activity/add-edit-activity-layout/EditActivityPage";
+import EditUserProfilePage from "./user/user-profile-layout/EditUserProfile";
+import UserRequestsPage from "./user/user-requests-layout/UserRequestsPage";
 import AdminRequestsPage from "./admin-requests-layout/AdminRequestsPage";
 import DeveloperPage from "./developers-layout/DeveloperPage";
 import UnauthorizedPage from "./UnauthorizedPage";
