+++ src/User/user-requests-layout/UserRequestsPage.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import apiClient from "./../../apiClient";
+import apiClient from "../../apiClient";
 import {Link, useNavigate} from "react-router-dom";
 import {handleError} from "./../../errorHandler";
